requirements.txt
dj-database-url==0.4.2
dj-static==0.0.6
Django==2.0
django-debug-toolbar==1.9.1
django-extensions==1.9.8
django-test-without-migrations==0.6
python-decouple==3.1
pytz==2017.3
six==1.11.0
sqlparse==0.2.4
static3==0.7.0
typing==3.6.2
codeclimate-test-reporter==0.2.3
djangorestframework==3.7.7
markdown==2.6.10
django-filter==1.1.0
